requirements.txt
aiogram==3.10.0
mistralai==0.1.2
redis==5.0.1
sqlalchemy==2.0.25
python-dotenv==1.0.0
aiohttp==3.9.3
beautifulsoup4==4.12.2
lxml==4.9.3
aiofiles==23.2.1
psycopg2-binary==2.9.9